fix: awaitable is required
